+++ injitest/src/main/java/io/mosip/test/mob/inji/pages/ScanPage.java
@@ -18,6 +18,12 @@ public class ScanPage extends BasePage{
 
     @AndroidFindBy(uiAutomator = "new UiSelector().textContains(\"Allow\")")
     private WebElement allowButton;
+    
+	@AndroidFindBy(className = "android.widget.ImageView")
+	private WebElement flipCamera;
+	
+	@AndroidFindBy(xpath = "//*[contains(@text,'Hold the phone steady and scan the QR code')]")
+	 private WebElement holdCameraSteady;
 
     public ScanPage(AppiumDriver driver) {
         super(driver);
@@ -32,4 +38,15 @@ public class ScanPage extends BasePage{
     public boolean isCameraOpen(){
     return isElementDisplayed(camera, "camera");
     }
+    
+    public boolean isCameraPageLoaded() {
+        return this.isElementEnabled(holdCameraSteady);
+    }
+	
+	
+	public boolean isFlipCameraClickable() {
+		return this.isElementEnabled(flipCamera);
+		
+	}
+
 }
